Refactored form components into individual files
